pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "fitz>=0.0.1.dev2",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "google-api-python-client>=2.166.0",
    "google-auth>=2.38.0",
    "openai>=1.72.0",
    "openpyxl>=3.1.5",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]

